+++ es6/client/template.js
@@ -13,6 +13,7 @@ const Template_ = ((mtr, $) => {
     // ------------------------------------------------------------------------
 
     let _isDev;
+    let _addBodyClassPrevSing;
 
     // ------------------------------------------------------------------------
     // Class Definition
@@ -83,6 +84,41 @@ const Template_ = ((mtr, $) => {
             }
         });
     }
+
+    // Get the parent template instance
+    Blaze.TemplateInstance.prototype.parentTemplate = function (levels) {
+        let view = this.view;
+        if (typeof levels === "undefined") {
+            levels = 1;
+        }
+        while (view) {
+            if (view.name.substring(0, 9) === "Template." && !(levels--)) {
+                return view.templateInstance();
+            }
+            view = view.parentView;
+        }
+    };
+
+    // Add body class
+    Blaze.addBodyClass = function (fn) {
+        if ($.isArray(fn)) {
+            return fn.forEach(Blaze.addBodyClass);
+        }
+        if (typeof fn !== 'function') {
+            $('body').removeClass(_addBodyClassPrevSing);
+            return mtr.startup(function () {
+                $('body').addClass(_addBodyClassPrevSing = fn);
+            });
+        }
+
+        mtr.startup(function () {
+            Tracker.autorun(function () {
+                $('body')
+                    .removeClass(fn._prev)
+                    .addClass(fn._prev = fn());
+            });
+        });
+    };
 
     // ------------------------------------------------------------------------
     // Meteor